add middleware

+++ pkg/repository/items.go
@@ -2,11 +2,11 @@ package repository
 
 import "github.com/bookmarks-api/models"
 
-func (r *Repository) GetAllItems() ([]models.Item, error) {
+func (r *Repository) GetAllItems(userId int) ([]models.Item, error) {
 	var items []models.Item
-	query := `select id, url, title from bookmarks_items`
+	query := `select id, url, title from bookmarks_items where user_id=$1`
 
-	err := r.db.Select(&items, query)
+	err := r.db.Select(&items, query, userId)
 	if err != nil {
 		return items, err
 	}
@@ -14,13 +14,13 @@ func (r *Repository) GetAllItems() ([]models.Item, error) {
 }
 
 func (r *Repository) AddItem(item *models.Item) error {
-	query := `insert into bookmarks_items (url, title) values($1, $2)`
-	_, err := r.db.Exec(query, item.Url, item.Title)
+	query := `insert into bookmarks_items (user_id, url, title) values($1, $2, $3)`
+	_, err := r.db.Exec(query, item.UserId, item.Url, item.Title)
 	return err
 }
 
-func (r *Repository) DeleteItem(id int) error {
-	query := `delete from bookmarks_items where id=$1`
-	_, err := r.db.Exec(query, id)
+func (r *Repository) DeleteItem(id, userId int) error {
+	query := `delete from bookmarks_items where id=$1 and user_id=$2`
+	_, err := r.db.Exec(query, id, userId)
 	return err
 }
